seq: Redesign steps_at_slot and play to be iterative.
Since they don't assume that the previous slot was the one that just
finished, this paves the way for a playhead that moves in different
directions.

Reshuffle the awkward first: and ending: parameters to play, in favor
of using prev_steps and adding a stop method, respectively.

Also thread a cycle count throughout to enable Step probability.