midi-utils: Clean up cc_fx_control_loop.
Now doing a lot of up-front parsing of the arguments, to prevent a
bunch of ad-hoc parsing later.

A range and quantum is now supported for callback mappings. The
default is 0..127 and 1.

Default quantum for fx mappings is now 0.01.

Split MIDI<->Range mapping logic into its own functions.

Initial CC messages for the default value are now always set, even if
no explicit default is given. In that case it uses the minimum of the
parameter's range.