seq: Automatically convert arguments to Tracks as needed.
It's now more sensible to feed things directly to Track.new, so do
that in mutator methods that take another track (e.g. zip). In the
case that a new Track needs to be constructed, it inherits the
granularity and timescale of the recipient, so that it will be pass
assert_compatible_track.