arp, seq: Fix Arp.arpeggiate sometimes not returning an array.
Particularly when passed the result of Sonic Pi's `chord` and using
certain simpler directions, arpeggiate would end up passing the Chord
object straight through. Even though Chord is a subclass of Array,
in-place mutations on it are pretty broken, so we need to be sure to
convert Chords to vanilla Arrays.