player, prob: Prepare for note_for_step to vary based on the owning Track.
In that case, we cannot call note_for_step on Steps that may have
come from another Track (e.g. the Steps in @prev_steps). Instead,
cache the resolved notes for @prev_steps in its own hash, and use
that in steps_at_slot.

Also requires a change in Prob - namely the (not_)prev_same_note
conditions cannot be evaluated based solely on Steps; they too must
be fed resolved notes from the Player - both for the steps in the
previous slot and for the step whose probability is being evaluated.